新增头像

+++ article/views.py
@@ -60,14 +60,19 @@ def create(request):
 
 # 删除文章
 def delete(request, id):
+    print(request.user)
     # 根据 id 获取需要删除的文章
     article = ArticlePost.objects.get(id=id)
+    # 过滤非作者的用户
+    if article.author != request.user:
+        return HttpResponse("抱歉，你无权删除这篇文章。")
     # 调用.delete()方法删除文章
     article.delete()
     # 完成删除后返回文章列表
     return redirect('article:article_list')
 
 
+# 更新文章
 def article_update(request, id):
     """
     更新文章的视图函数
@@ -75,8 +80,12 @@ def article_update(request, id):
     GET方法进入初始表单页面
     id： 文章的 id
     """
+    print(request.user)
     # 获取需要修改的具体文章对象
     article = ArticlePost.objects.get(id=id)
+    # 过滤非作者的用户
+    if article.author != request.user:
+        return HttpResponse("抱歉，你无权修改这篇文章。")
     # 判断用户是否为 POST 提交表单数据
     if request.method == 'POST':
         # 将提交的数据赋值到表单实例中
